+++ src/main/java/slogo/InstructionClasses/Power.java
@@ -9,6 +9,6 @@ public class Power extends MathTwoParam{
 
   @Override
   public String toString() {
-    return String.format("power %d %d\n", (int) getParamOne(), (int) getParamTwo());
+    return String.format("power %f %f\n", getParamOne(), getParamTwo());
   }
 }
